+++ backend/windmill-api/openapi-deref.yaml
@@ -1,6 +1,6 @@
 openapi: 3.0.3
 info:
-  version: 1.246.4
+  version: 1.246.15
   title: Windmill API
   contact:
     name: Windmill Team
