+++ docker-compose.yml
@@ -11,6 +11,8 @@ services:
       - db_data:/var/lib/postgresql/data
     expose:
       - 5432
+    ports:
+      - 5432:5432
     environment:
       POSTGRES_PASSWORD: changeme
       POSTGRES_DB: windmill
